Fix upload

+++ index.js
@@ -102,7 +102,7 @@ function StreamUpload (options) {
     const __checkFileType = function (type) {
         if (that.settings.allowedTypes && that.settings.allowedTypes.length) {
             for (let i = 0; i < that.settings.allowedTypes.length; i++) {
-                const exp = new RegExp(that.settings.allowedTypes[i]);
+                const exp = new RegExp(that.settings.allowedTypes[i].replace('+','\\+').replace('.','\\.'));
                 if (exp.test(type)) {
                     return true;
                 }
@@ -181,7 +181,9 @@ function StreamUpload (options) {
             s3
                 .deleteObject({Key: that.filename})
         } else {
+          if (fs.existsSync(that.filename)) {
             fs.unlink(that.filename);
+          }
         }
     }
